+++ packages/icon/src/types/icon.types.ts
@@ -1,19 +1,12 @@
 import { ComponentSize } from "@impulse-ui-native/theme";
 import { ComponentType } from "react";
-import { ViewStyle } from "react-native";
+import { SvgProps } from "react-native-svg";
 
-export interface IconStyle extends ViewStyle {
-  color?: string;
-}
-
-export interface IconSvgProps {
-  width: number;
-  height: number;
-  color?: string;
-}
+export type IconStyle = SvgProps["style"];
 
 export interface IconProps {
   size?: ComponentSize | number;
   color?: string;
-  icon: ComponentType<IconSvgProps>;
+  icon: ComponentType<SvgProps>;
+  style?: IconStyle;
 }