+++ packages/theme/src/hooks/use-colors.hook.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 import { useThemeContext } from "../providers";
-import { ThemeColors } from "../theme/light.theme";
+import { ThemeColors } from "../types";
 
 export function useColors(): ThemeColors {
   const context = useThemeContext();