+++ packages/theme/src/theme/index.ts
@@ -1,3 +1,2 @@
-export { LightTheme } from './light.theme';
-export { DarkTheme } from './dark.theme';
-export type { AppTheme } from './light.theme';
+export { DarkTheme } from "./dark.theme";
+export { LightTheme } from "./light.theme";