+++ packages/theme/src/types/theme-provider.types.ts
@@ -1,4 +1,12 @@
-import { AppTheme } from "../theme";
+import { DeepPartial } from "@impulse-ui-native/core";
+import {
+  BorderSizes,
+  FontSizes,
+  NeutralColors,
+  Radii,
+  Space,
+} from "./tokens.types";
+import { TypographyPresetKey, TypographyProps } from "./typography.types";
 
 export interface ThemeContextData {
   loaded: boolean;
@@ -6,3 +14,42 @@ export interface ThemeContextData {
 }
 
 export type ColorScheme = "light" | "dark";
+
+export type ThemeColors = {
+  primary: string;
+  secondary: string;
+  neutral: NeutralColors;
+
+  white: string;
+  black: string;
+
+  error: string;
+  warning: string;
+  success: string;
+  info: string;
+
+  text: {
+    primary: string;
+    secondary: string;
+  };
+
+  background: {
+    primary: string;
+    secondary: string;
+  };
+} & ExtendedThemeColors;
+
+export interface ExtendedThemeColors {}
+
+export interface ComponentsThemes {
+  typography: Record<TypographyPresetKey, TypographyProps>;
+}
+
+export interface AppTheme {
+  colors: ThemeColors;
+  spacing: Space;
+  radii: Radii;
+  borderSize: BorderSizes;
+  fontSize: FontSizes;
+  components?: DeepPartial<ComponentsThemes>;
+}