+++ packages/typography/src/constants/typography.constants.ts
@@ -1,0 +1,29 @@
+import {
+  FontWeightItalic,
+  FontWeightKey,
+  FontWeightNormal,
+} from "@impulse-ui-native/theme";
+
+export const FontWeightsNormal: Record<FontWeightKey, FontWeightNormal> = {
+  100: "Montserrat_100Thin",
+  200: "Montserrat_200ExtraLight",
+  300: "Montserrat_300Light",
+  400: "Montserrat_400Regular",
+  500: "Montserrat_500Medium",
+  600: "Montserrat_600SemiBold",
+  700: "Montserrat_700Bold",
+  800: "Montserrat_800ExtraBold",
+  900: "Montserrat_900Black",
+};
+
+export const FontWeightsItalic: Record<FontWeightKey, FontWeightItalic> = {
+  100: "Montserrat_100Thin_Italic",
+  200: "Montserrat_200ExtraLight_Italic",
+  300: "Montserrat_300Light_Italic",
+  400: "Montserrat_400Regular_Italic",
+  500: "Montserrat_500Medium_Italic",
+  600: "Montserrat_600SemiBold_Italic",
+  700: "Montserrat_700Bold_Italic",
+  800: "Montserrat_800ExtraBold_Italic",
+  900: "Montserrat_900Black_Italic",
+};