+++ packages/core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @impulse-ui-native/core
 
+## 1.0.6
+
+### Patch Changes
+
+- Exposed new props for Typography, Button and Theme interfaces
+
+## 1.0.5
+
+### Patch Changes
+
+- Adding Typography and View pacakges to the toolkit
+
+## 1.0.4
+
+### Patch Changes
+
+- Fixing icon component
+
+## 1.0.3
+
+### Patch Changes
+
+- Updating icon component
+
 ## 1.0.2
 
 ### Patch Changes