+++ packages/safe-area-view/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @impulse-ui-native/safe-area-view
 
+## 1.0.6
+
+### Patch Changes
+
+- Exposed new props for Typography, Button and Theme interfaces
+- Updated dependencies
+  - @impulse-ui-native/theme@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- Adding Typography and View pacakges to the toolkit
+- Updated dependencies
+  - @impulse-ui-native/theme@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Fixing icon component
+- Updated dependencies
+  - @impulse-ui-native/theme@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Updating icon component
+- Updated dependencies
+  - @impulse-ui-native/theme@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes