+++ packages/toolkit/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @impulse-ui-native/toolkit
 
+## 1.0.6
+
+### Patch Changes
+
+- Exposed new props for Typography, Button and Theme interfaces
+- Updated dependencies
+  - @impulse-ui-native/safe-area-view@1.0.6
+  - @impulse-ui-native/layer-manager@1.0.6
+  - @impulse-ui-native/typography@1.0.6
+  - @impulse-ui-native/button@1.0.6
+  - @impulse-ui-native/flyout@1.0.6
+  - @impulse-ui-native/input@1.0.6
+  - @impulse-ui-native/theme@1.0.6
+  - @impulse-ui-native/core@1.0.6
+  - @impulse-ui-native/echo@1.0.6
+  - @impulse-ui-native/icon@1.0.6
+  - @impulse-ui-native/view@1.0.6
+  - @impulse-ui-native/tag@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- Adding Typography and View pacakges to the toolkit
+- Updated dependencies
+  - @impulse-ui-native/safe-area-view@1.0.5
+  - @impulse-ui-native/layer-manager@1.0.5
+  - @impulse-ui-native/typography@1.0.5
+  - @impulse-ui-native/button@1.0.5
+  - @impulse-ui-native/flyout@1.0.5
+  - @impulse-ui-native/input@1.0.5
+  - @impulse-ui-native/theme@1.0.5
+  - @impulse-ui-native/core@1.0.5
+  - @impulse-ui-native/echo@1.0.5
+  - @impulse-ui-native/icon@1.0.5
+  - @impulse-ui-native/view@1.0.5
+  - @impulse-ui-native/tag@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Fixing icon component
+- Updated dependencies
+  - @impulse-ui-native/safe-area-view@1.0.4
+  - @impulse-ui-native/layer-manager@1.0.4
+  - @impulse-ui-native/button@1.0.4
+  - @impulse-ui-native/flyout@1.0.4
+  - @impulse-ui-native/input@1.0.4
+  - @impulse-ui-native/theme@1.0.4
+  - @impulse-ui-native/core@1.0.4
+  - @impulse-ui-native/echo@1.0.4
+  - @impulse-ui-native/icon@1.0.4
+  - @impulse-ui-native/tag@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Updating icon component
+- Updated dependencies
+  - @impulse-ui-native/button@1.0.3
+  - @impulse-ui-native/core@1.0.3
+  - @impulse-ui-native/echo@1.0.3
+  - @impulse-ui-native/flyout@1.0.3
+  - @impulse-ui-native/icon@1.0.3
+  - @impulse-ui-native/input@1.0.3
+  - @impulse-ui-native/layer-manager@1.0.3
+  - @impulse-ui-native/safe-area-view@1.0.3
+  - @impulse-ui-native/tag@1.0.3
+  - @impulse-ui-native/theme@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes